extract day count data fetching logic
Extract useDayCounts hook from DateRangeFilterGroup component to reduce component complexity and enable reuse. Simplify component logic by inlining simple helper functions.

Change-Type: refactor
Scope: frontend
